Refactoring of multiple field operators stencils (#47)
Re-factoring of multiple field operators:

* mo_nh_diffusion_stencil_16, mo_nh_diffusion_stencil_08, mo_intp_rbf_rbf_vec_interpol_vertex_dom_b

* mo_solve_nonhydro_stencil_07, mo_solve_nonhydro_stencil_13

* mo_solve_nonhydro_stencil_14

* mo_solve_nonhydro_stencil_32, mo_solve_nonhydro_stencil_33

* mo_solve_nonhydro_stencil_41, mo_solve_nonhydro_stencil_42

* mo_solve_nonhydro_stencil_47, mo_solve_nonhydro_stencil_50

+++ atm_dyn_iconam/tests/test_mo_solve_nonhydro_stencil_26.py
@@ -33,7 +33,7 @@ def test_mo_solve_nonhydro_stencil_26():
 
     z_graddiv_vn = random_field(mesh, EdgeDim, KDim)
     vn = random_field(mesh, EdgeDim, KDim)
-    scal_divdamp_o2 = np.float64(5.0)
+    scal_divdamp_o2 = 5.0
 
     ref = mo_solve_nonhydro_stencil_26_numpy(
         np.asarray(z_graddiv_vn), np.asarray(vn), scal_divdamp_o2
